feat(api): wire end-to-end scoring into POST /transactions (Phase 3C-2)
New scoring service at app/services/scoring.py loads the TransactionContext once (customer + merchant + 180 days of recent history in one indexed query), runs the six-rule engine, calls the cached XGBoost+SHAP explainer, and composes the final decision per the conservative-wins matrix in PHASE_3_DESIGN.md. HARD_BLOCK rules short-circuit the model and return DECLINE; REVIEW rules promote APPROVE to REVIEW but never override a model DECLINE. The service does not commit the session — idempotency.store commits Transaction + audit_log + idempotency_keys atomically, so a replayed key returns the byte-for-byte original response.

Every scoring call writes one audit_log row with actor=scorer:{trained_at_utc} (e.g. scorer:2026-05-21T12:47:25+00:00), giving every decision a permanent trace to the model version that produced it. The trained_at_utc value is read once at module import from ml/artifacts/training_metadata.json, falling back to 'unknown' if the file is absent so the test suite imports cleanly in artifact-free environments.

Two latent bugs surfaced during integration and were fixed in this commit: (1) rules.py compared offset-aware in-memory datetimes against offset-naive DB-loaded datetimes, raising TypeError on the first real POST — fixed with a local _ensure_utc helper at every created_at comparison, mirroring the Phase 3C-1 pattern in features.py; (2) audit_log.id was declared as BigInteger but SQLite only auto-increments PRIMARY KEY columns of declared type INTEGER, so INSERTs failed with NOT NULL constraint violations — fixed with a new Alembic migration (c3d4e5f6a7b8) altering the column type, non-destructive because the table was empty.

New ml/benchmark_latency.py measures both layers separately. Service-layer (score_transaction called directly): p50=3.7ms, p95=5.8ms, p99=8.9ms across n=500. Endpoint (full HTTP round-trip via httpx): p50=16ms, p95=20ms. p99 omitted from the README headline because developer-laptop SQLite shows intermittent ~100ms file-lock variance unrelated to scoring; full distribution including p99 is committed to ml/artifacts/latency_metrics.json. README updated with the measured numbers and a methodology blockquote.

Tests: 10 new unit tests in test_scoring.py (decision matrix parametrized over 5 cases, hard-block short-circuit, review-with-approve-or-decline composition, audit-on/off, no-commit invariant) plus 4 from the matrix parametrize — 14 total. 6 new integration tests in test_scoring_endpoint.py exercise the full POST flow with the real explainer (clean transaction, velocity burst, high amount, idempotent replay, top contributors, audit log persistence). The 2 Phase 3B PENDING tests in test_transactions_endpoint.py were renamed and tightened to assert real APPROVE/REVIEW/DECLINE values. 170 tests passing in the full suite (150 prior + 14 scoring unit + 6 scoring integration).

+++ backend/app/models/audit_log.py
@@ -1,5 +1,5 @@
 """Audit log ORM model — append-only record of all state changes."""
-from sqlalchemy import BigInteger, Index, String, Text
+from sqlalchemy import Index, Integer, String, Text
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.models.base import Base, TimestampMixin
@@ -16,7 +16,7 @@ class AuditLog(Base, TimestampMixin):
 
     __tablename__ = "audit_log"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     actor: Mapped[str] = mapped_column(String(128), nullable=False)
     action: Mapped[str] = mapped_column(String(64), nullable=False)
     resource_type: Mapped[str] = mapped_column(String(64), nullable=False)
